move out config to modules
